+++ src/misc/mlx5dvsymbols.cc
@@ -1,3 +1,6 @@
+// Modification Copyright (c) Advanced Micro Devices, Inc., or its affiliates.
+// SPDX-License-Identifier: MIT
+
 #include <sys/types.h>
 #include <unistd.h>